Missed some refactoring there
