+++ ppb/features/twophase.py
@@ -2,7 +2,7 @@
 A system for two phase updates: Update, and Commit.
 """
 from dataclasses import dataclass
-from ppb.systems import System
+from ppb.systemslib import System
 from ppb.eventlib import EventMixin
 
 __all__ = 'Commit',
